backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
pypdf==6.5.0
python-dotenv==1.2.1
requests==2.31.0
sqlparse==0.5.5
